requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
requests==2.32.3
python-dotenv==1.0.1
pandas>=2.0.0
pdfplumber==0.11.0
python-multipart==0.0.9
sqlmodel==0.0.22
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
